# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/betSlip/getSequenceNumber": {
            "get": {
                "tags": [
                    "BetSlip"
                ],
                "summary": "取得單一注單將號2.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/betSlip/getSequenceNumbers": {
            "get": {
                "tags": [
                    "BetSlip"
                ],
                "summary": "取得複數注單將號2.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "數量(可接受範圍1-50)",
                        "name": "quantity",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/betSlip/roundCheck": {
            "get": {
                "tags": [
                    "BetSlip"
                ],
                "summary": "取得需補注單列表2.3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "開始時間",
                        "name": "fromDate",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "結束時間",
                        "name": "toDate",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/betSlipPersonal/addUniversalGameLog": {
            "post": {
                "tags": [
                    "BetSlipPersonal"
                ],
                "summary": "寫入遊戲紀錄3.3",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.GameLog"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
        "/api/gameReport/orderList": {
            "get": {
                "tags": [
                    "GameReport"
                ],
                "summary": "取得歷史紀錄網址4.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "連線令牌",
                        "name": "connectToken",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1.0/activity/ranking/distribution": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "活動派獎6.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Distribution"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1.0/activity/ranking/settlement": {
            "post": {
                "tags": [
                    "Activity"
                ],
                "summary": "活動結算6.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Settlement"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1.0/betSlipPersonal/gameResult": {
            "post": {
                "tags": [
                    "BetSlipPersonal"
                ],
                "summary": "寫入賽果(拉霸)3.1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.GameResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
        "/api/v1.0/betSlipPersonal/supplement/result": {
            "post": {
                "tags": [
                    "BetSlipPersonal"
                ],
                "summary": "補單補賽果單3.2",
                "parameters": [
                    {
                        "type": "string",
                        "description": "auth token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.FinishGameResult"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.BaseHttpResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.BaseHttpResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "status": {
                    "$ref": "#/definitions/entity.HttpResponseStatus"
                }
            }
        },
...
        "entity.Distribution": {
            "type": "object",
            "properties": {
                "activityIV": {
                    "type": "string",
                    "minLength": 1
                },
                "prizePayout": {
                    "type": "number"
                },
                "rank": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "entity.FinishGameResult": {
            "type": "object",
            "properties": {
                "betMode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ]
                },
                "betTime": {
                    "type": "string"
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "currencyKindBet": {
                    "type": "number"
                },
                "currencyKindContribution": {
                    "type": "number"
                },
                "currencyKindJackPot": {
                    "type": "number"
                },
                "currencyKindPayout": {
                    "type": "number"
                },
                "currencyKindWinLose": {
                    "type": "number"
                },
                "freeGame": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "gameRoom": {
                    "type": "integer",
                    "minimum": 1
                },
                "gameSequenceNumber": {
                    "type": "string",
                    "minLength": 1
                },
                "sequenceID": {
                    "type": "string",
                    "minLength": 1
                },
                "serverTime": {
                    "type": "string"
                },
                "transID": {
                    "type": "string"
                },
                "turnTimes": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "entity.GameLog": {
            "type": "object",
            "properties": {
                "bet": {
                    "type": "number"
                },
                "betTime": {
                    "type": "string"
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "contribution": {
                    "type": "number"
                },
                "currencyKindBet": {
                    "type": "number"
                },
                "currencyKindContribution": {
                    "type": "number"
                },
                "currencyKindJackPot": {
                    "type": "number"
                },
                "currencyKindPayout": {
                    "type": "number"
                },
                "currencyKindWinLose": {
                    "type": "number"
                },
                "gameLog": {},
                "gameSequenceNumber": {
                    "type": "string",
                    "minLength": 1
                },
                "jackPot": {
                    "type": "number"
                },
                "payout": {
                    "type": "number"
                },
                "sequenceID": {
                    "type": "string",
                    "minLength": 1
                },
                "winLose": {
                    "type": "number"
                }
            }
        },
        "entity.GameResult": {
            "type": "object",
            "properties": {
                "betMode": {
                    "type": "integer",
                    "enum": [
                        0,
                        1,
                        2,
                        3
                    ]
                },
                "betTime": {
                    "type": "string"
                },
                "connectToken": {
                    "type": "string",
                    "minLength": 1
                },
                "currencyKindBet": {
                    "type": "number"
                },
                "currencyKindContribution": {
                    "type": "number"
                },
                "currencyKindJackPot": {
                    "type": "number"
                },
                "currencyKindPayout": {
                    "type": "number"
                },
                "currencyKindWinLose": {
                    "type": "number"
                },
                "freeGame": {
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "gameRoom": {
                    "type": "integer",
                    "minimum": 1
                },
                "gameSequenceNumber": {
                    "type": "string",
                    "minLength": 1
                },
                "sequenceID": {
                    "type": "string",
                    "minLength": 1
                },
                "serverTime": {
                    "type": "string"
                },
                "turnTimes": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "entity.HttpResponseStatus": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "dateTime": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "traceCode": {
                    "type": "string"
                }
            }
        },
...
        "entity.Settlement": {
            "type": "object",
            "properties": {
                "activityIV": {
                    "type": "string",
                    "minLength": 1
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "CNY",
                        "JPY",
                        "THB",
                        "MMK",
                        "VND",
                        "MYR",
                        "IDR",
                        "USD",
                        "USDT",
                        "BDT",
                        "PHP",
                        "kVND",
                        "kIDR",
                        "SGD",
                        "KRW",
                        "INR",
                        "HKD"
                    ]
                },
                "gameSequenceNumber": {
                    "type": "string"
                },
                "memberID": {
                    "type": "integer"
                },
                "prize": {
                    "type": "number"
                },
                "rank": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
...
    }
}